fix some errors for esp32h2 platform

+++ src/platform/OpenThread/GenericThreadStackManagerImpl_OpenThread.cpp
@@ -846,7 +846,7 @@ CHIP_ERROR GenericThreadStackManagerImpl_OpenThread<ImplClass>::_WriteThreadNetw
 
     case ThreadNetworkDiagnostics::Attributes::Ids::RoutingRole: {
         otDeviceRole role = otThreadGetDeviceRole(mOTInst);
-        err               = encoder.Encode(role);
+        err               = encoder.Encode(static_cast<uint8_t>(role));
     }
     break;
